pyproject.toml
[project]
name = "travel-mapper"
version = "0.1.0"
description = "Tool for mapping tripit data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "argon2-cffi-bindings>=21.2.0",
  "argon2-cffi>=21.3.0",
  "certifi>=2023.7.22",
  "cffi>=1.15.1",
  "fastapi[standard]>=0.115.8",
  "opentelemetry-api==1.29",
  "opentelemetry-instrumentation-sqlalchemy",
  "opentelemetry-sdk==1.29",
  "orjson>=3.10.15",
  "pydantic-settings>=2.8.0",
  "pydantic>=2.10.6",
  "pyjwt>=2.7.0",
  "pytest-asyncio>=0.21.0",
  "pytest>=7.3.2",
  "python-dotenv>=1.0.0",
  "requests>=2.32.3",
  "sqlalchemy[asyncio]>=2.0.38",
  "uvicorn[standard]>=0.34.0",
  "requests-oauthlib>=2.0.0",
]

[project.scripts]
create_db = "create_db.py"

[tool.uv]
dev-dependencies = ["fastapi-cli>=0.0.7", "ruff>=0.9.6"]
